PackageXml2 dropped anon namespace
